feat: Add userName field #115
